// src/apis/generated/model/getEnterpriseQuantsInfoResponseDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * BFF API
 * Frontend 向け BFF API
 * OpenAPI spec version: 1.0
 */
import type { EnterpriseQuantInfoDto } from './enterpriseQuantInfoDto';

export interface GetEnterpriseQuantsInfoResponseDto {
  /** 企業別クオンツ情報一覧 */
  enterprises: EnterpriseQuantInfoDto[];
  /** レスポンス更新日時 */
  updatedAt: string;
  /** データ基準日 */
  dataAsOfDate: string;
  /** リアルタイムデータかどうか */
  isRealtime: boolean;
  /** 利用上の注意 */
  disclaimers: string[];
}